fixed publish when edit issue. Added anon feature to language pair leaderboard
